+++ test/org/capitalism/TransactionTest.java
@@ -1,7 +1,5 @@
 package org.capitalism;
 
-import java.math.BigDecimal;
-
 import junit.framework.Assert;
 
 import org.junit.Test;
@@ -16,7 +14,7 @@ public class TransactionTest {
 		
 		seller.createConsumable();
 		
-		final BigDecimal totalMoney = Money.MoneySupply.getTotalMoney();
+		final double totalMoney = Money.MoneySupply.getTotalMoney();
 		final double buyerWorth = buyer.getWorth();
 		final double sellerWorth = seller.getWorth();
 		final int numSellerConsumables = seller.getNumConsumables();
@@ -31,7 +29,7 @@ public class TransactionTest {
 		
 		new Transaction(buyer, seller, agreement);
 		
-		final BigDecimal totalMoneyAfterTrans = Money.MoneySupply.getTotalMoney();
+		final double totalMoneyAfterTrans = Money.MoneySupply.getTotalMoney();
 		final double buyerPostWorth = buyer.getWorth();
 		final double sellerPostWorth = seller.getWorth();
